backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI and Language Model dependencies
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6

# Optional dependencies for enhanced features
pymongo>=4.3.3  # Optional - for MongoDB integration
openai>=1.3.0   # Optional - for vector embeddings
requests>=2.31.0  # For external API calls

